// frontend/src/rebuild/learning/horizonte/space1/polyhedra.generated.ts
export const SECTION_SOLIDS = ['tetrahedron', 'cube', 'octahedron', 'cylinder'] as const;
export type SectionSolid = (typeof SECTION_SOLIDS)[number];
// ...
export const SHAPES = ['triangle', 'square', 'rectangle', 'rhombus', 'parallelogram', 'trapezoid', 'quadrilateral', 'pentagon', 'hexagon', 'polygon', 'circle', 'ellipse'] as const;
export type SectionShape = (typeof SHAPES)[number];
// ...
export interface Plane { normal: readonly [number, number, number]; offset: number }